requirements.txt
streamlit
openai
faiss-cpu
requests
authlib
numpy
